requirements.txt
asgiref==3.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-environ==0.8.1
django-js-asset==2.0.0
django-mptt==0.13.4
gunicorn==20.1.0
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
